enrich school information

+++ src/main/java/com/tangenta/gkassist/school/representation/SchoolBriefRepresentation.java
@@ -6,17 +6,17 @@ import com.tangenta.gkassist.school.model.SchoolId;
 import java.util.List;
 
 public class SchoolBriefRepresentation implements Representation {
-    private int totalNum;
+    private long totalNum;
     private List<SchoolId> schoolIds;
     private List<String> schoolNames;
 
-    public SchoolBriefRepresentation(int totalNum, List<SchoolId> schoolIds, List<String> schoolNames) {
+    public SchoolBriefRepresentation(long totalNum, List<SchoolId> schoolIds, List<String> schoolNames) {
         this.totalNum = totalNum;
         this.schoolIds = schoolIds;
         this.schoolNames = schoolNames;
     }
 
-    public int getTotalNum() {
+    public long getTotalNum() {
         return totalNum;
     }
